Do not patch field managers for Patch resources (#2640)
### Proposed changes

This pull request addresses the issue where Patch resources
unintentionally patch field managers during update operations (ie. when
a Patch resource is updated and reused). Such modification results in
the Patch resource taking control of all fields managed owned by a field
manager with the prefix `kubectl`. This, in turn, leads to unintended
unset fields when running `pulumi up`. Patch resources are designed to
contain a specific subset of fields for updating, and they should not
interfere with the field managers of normal resources. It's important to
note that field managers still need to be patched for regular resources,
as they facilitate the transition from Client-Side Apply (CSA) to
Server-Side Apply (SSA) for existing Pulumi-managed resources.

Additionally, this PR includes a new test to validate this behavior. The
test will fail if the logic to avoid field manager patching for Patch
resources is not implemented.

### Related issues:
Fixes: #2639