Add Kustomize test
The invoke used for Kustomize was already fixed to be resilient to `None` or empty dicts being returned in a907bddce8da5141992c2c9946e2f6be3fe9f79a. This test, which is based on the existing `kustomize` test is to ensure it continues to work without an error being raised.