Fix: Make the invoke calls for Helm charts and YAML config resilient to the value being `None` or an empty dict (#2665)
### Proposed changes

In v3.92.0 of the core `pulumi` Python SDK, invokes were changed to
return an empty dict rather than `None` for empty results, which is
consistent with other language SDKs and the original behavior of the
Python SDK, and makes generated provider Python SDKs work with empty
results. However, the hand-rolled invoke code in the Kubernetes Python
SDK for Helm and YAML are not resilient to it getting an empty dict.
This change fixes Helm and YAML to be resilient, in the same way that
the invoke for Kustomize was updated in
a907bddce8da5141992c2c9946e2f6be3fe9f79a.

An invoke will return empty results when called on an unconfigured
provider (i.e. a provider whose configuration contains an unknown
value). The first three commits add regression tests that pass an
unknown value as a provider's configuration. The Helm and YAML
regression tests fail before the fix and succeed after the fix. The test
for Kustomize already succeeds because of the change in
a907bddce8da5141992c2c9946e2f6be3fe9f79a. The new tests are based on
existing tests.

To ensure older versions of `pulumi-kubernetes` continue to work with
newer versions of `pulumi`, we're planning to patch the `pulumi` SDK to
return `None` for empty invoke returns when the function is one of the
affected kubernetes function tokens. This is tracked by
https://github.com/pulumi/pulumi/issues/14508 (PR:
https://github.com/pulumi/pulumi/pull/14535).

### Related issues

Fixes #2664