Fix panic when PULUMI_KUBERNETES_MANAGED_BY_LABEL is used on resources without the managed by label (#2711)
### Proposed changes

This pull request addresses a panic scenario that occurs when attempting
to assert a nil object as a string. The issue arises specifically when
the `app.kubernetes.io/managed-by` label is absent from the Kubernetes
object. This is the case when the object is created using SSA mode. The
fix ensures proper handling of this situation to prevent the panic.

In addition to the fix, a new unit test has been introduced to confirm
the resolution of this panic scenario.

### Related issues (optional)

Fixes: #2708