Fix resource kind lookup using URN.Type instead of QualifiedType (#2719)
This pull request uses the URN`Type` method, rather than `QualifiedType`
to search if resources are a patch or list resource. Qualified types
include the parent resource if there is one, so it would have failed the
lookup.

Additionally, comprehensive unit tests have been included to validate
this change.

Fixes: #2718 (manual validation was done to ensure this fix solves this
CUJ)