Fix: DiffConfig fails when provider's kubeconfig is set to file path (#2771)
<!--Thanks for your contribution. See [CONTRIBUTING](CONTRIBUTING.md)
    for Pulumi's contribution guidelines.

    Help us merge your changes more quickly by adding more details such
    as labels, milestones, and reviewers.-->

### Proposed changes

<!--Give us a brief description of what you've done and what it solves.
-->

This PR refactors the kubeconfig loading code to be more consistent
between `DiffConfig` and `Configure`. Previously the `DiffConfig` method
did not support file-based kubeconfigs, and would throw a spurious
parser error.

`DiffConfig` loads the kubeconfig to be able to detect cluster
replacement, by comparing the active cluster info. Unlike `Configure`,
it doesn't have support for ambient kubeconfigs. When an ambient config
is in use, or the config is invalid (e.g. the `context` or `cluster`
properties refer to a non-existent entry), then the replacement logic is
skipped.

This PR also fixes a panic in `DiffConfig` that would occur if `cluster`
or `context` is a computed value; the provider will conservatively
suggest replacement as it does when `kubeconfig` is a computed value.

### Testing
New test cases were added for cluster change detection.

New tests were added for the parsing code that was consolidated into
`utils.go`. The tests now cover file-, string-, and object-based
resource property values for `kubeconfig`.

### Related issues (optional)

<!--Refer to related PRs or issues: #1234, or 'Fixes #1234' or 'Closes
#1234'.
Or link to full URLs to issues or pull requests in other GitHub
repositories. -->
Closes: #2663 
Related: #1032