skip normalization in preview w/ computed fields (#2846)
<!--Thanks for your contribution. See [CONTRIBUTING](CONTRIBUTING.md)
    for Pulumi's contribution guidelines.

    Help us merge your changes more quickly by adding more details such
    as labels, milestones, and reviewers.-->

### Proposed changes

<!--Give us a brief description of what you've done and what it solves.
-->

Skips the normalization step in the presence of computed inputs. 

Note: previously when `clients.Normalize` was called with a `Secret`, it
would panic. If called with a non-secret, an internal error would be
masked and the original object would be used.

### Related issues (optional)

<!--Refer to related PRs or issues: #1234, or 'Fixes #1234' or 'Closes
#1234'.
Or link to full URLs to issues or pull requests in other GitHub
repositories. -->
Closes #2845