ConfigFile V2 (Component) (#2862)
This PR implements a new `ConfigFile` resource (v2) as a Pulumi
multi-language component (MLC). The new resource is offered side-by-side
with the existing (v1) resource; it is not a drop-in replacement at this
time.

Some notable differences with respect to the previous (overlay)
implementations:
1. Implemented as an MLC to have a consistent implementation across SDKs
and to extend support for YAML and Java.
2. The component name is used as a prefix for the child resource names.
Use the `resourcePrefix` property to override. Note that the
`resourcePrefix` is not applied to the Kubernetes object names.
4. Transformations aren't supported (yet). Support for Pulumi-style
transformation is being tracked
[here](https://github.com/pulumi/pulumi/issues/12996), and
Kubernetes-style transformation may come in future.
5. The `file` property is required and the file must exist.

Closes #2785