Upgrade postgres operator integration test to use v1.11.0 (#2891)
The v1.7.0 version of the Zolando Posgres Operator chart is no longer
being hosted causing our CI tests to fail (eg.
https://github.com/pulumi/pulumi-kubernetes/actions/runs/8300755190/job/22843283137).
This PR bumps the chart version to a more recent 1.11.0. I've inspected
the new version to ensure that the intent of the test isn't lost by
upgrading to this version.