Fix: Missing Kind in preview mode (yaml/v2) (#2904)
<!--Thanks for your contribution. See [CONTRIBUTING](CONTRIBUTING.md)
    for Pulumi's contribution guidelines.

    Help us merge your changes more quickly by adding more details such
    as labels, milestones, and reviewers.-->

### Proposed changes

<!--Give us a brief description of what you've done and what it solves.
-->
The yaml/v2 package eagerly resolves kinds to apply a default namespace.
This PR addresses a problem with the use case
where a kind is installed by one `ConfigGroup` and then used by another.
In preview mode, the kind cannot be found by the latter group because
the expected side-effect (i.e. installation of a CRD) doesn't occur.

The solution to the issue is to maintain a cache in the provider of new
CRDs. When the provider executes `Check` on a `CustomResourceDefinition`
resource, it adds the CRD to the cache. Later, when a component resource
must resolve a given kind, it checks the cache. In this way, the cache
compensates for the lack of side-effects (i.e. API server calls) that
would normally allow the kind to be resolved. Overall ordering is
established using `DependsOn`.

It is reasonable to populate the cache during `Check` because a)`Check`
produces the "planned" state that is needed later without actuating it,
and b) `Update` is called conditionally and thus isn't a reliable
alternative.

### Implementation Details
The provider has three overlapping modes of operation - preview-mode,
`clusterUnreachable` and `yamlRenderMode` - that affect the
initialization and use of `clientSet`. Previously when
`clusterUnreachable` is true then the `clientSet` is left null. Now, we
new-up the `clientSet` itself but leave its various fields null. This
change makes it possible to use the CRD cache in all modes, as was
necessary to support `yamlRenderMode`.

### Testing
A new integration test is added called `yamlv2` that exercises this
scenario.

Manual testing of "cluster unreachable mode" was done by misconfiguring
my local environment (to have an invalid kube context). Tested in
combination with "yaml rendering mode" and found that the latter works
as expected even when the cluster is unreachable.

### Related issues (optional)

<!--Refer to related PRs or issues: #1234, or 'Fixes #1234' or 'Closes
#1234'.
Or link to full URLs to issues or pull requests in other GitHub
repositories. -->