[dotnet] Unknowns for previews involving an uninitialized provider (#2957)
<!--Thanks for your contribution. See [CONTRIBUTING](CONTRIBUTING.md)
    for Pulumi's contribution guidelines.

    Help us merge your changes more quickly by adding more details such
    as labels, milestones, and reviewers.-->

### Proposed changes

This PR fixes the "panic" that occurs when an uninitialized provider is
used with the yaml/kustomize resources. In that situation, the invoke
calls aren't performed and the engine simply returns an uninitialized
result. This causes a problem in the callers because the methods on
`ImmutableArray` mustn't be used when the value is uninitialized
(`IsDefault` is true).

### Example
The example given in #2741 easily repros the issue. With this fix, the
behavior improves:

```
Diagnostics:
  kubernetes:yaml:ConfigFile (guestbook):
    warning: Required input properties have unknown values. Preview is incomplete.

Outputs:
    FrontendIp: output<string>
```

### Testing
Includes integration tests involving an uninitialized provider with a
`ConfigFile` or `Directory`.

### Related issues (optional)

<!--Refer to related PRs or issues: #1234, or 'Fixes #1234' or 'Closes
#1234'.
Or link to full URLs to issues or pull requests in other GitHub
repositories. -->

Closes #2741