Fix a couple tests not running on CI (#2982)
I noticed as part of
https://github.com/pulumi/pulumi-kubernetes/pull/2953 that the
`tests/convert` and `tests/provider` directories aren't currently
getting exercised in CI, and as a result one of them is failing on
master.

This PR moves `tests/provider` and `tests/convert` under `sdk/java`,
which is currently executed in CI but doesn't contain any tests. We also
fix the convert test along the way.

"sdk/java" is admittedly an odd namespace to put these under, but our CI
targets don't give us much flexibility at the moment, and at the end of
the day it's just an arbitrary way of sharding things.