Fix a panic when diffing nil values against missing keys (#2987)
Users can run into a panic because we invoke `addPatchValueToDiff` like
so:

```go
pc.addPatchValueToDiff(append(path, k), v, old[k], newInput[k], oldInput[k], inArray)
```

These are `map[string]any` types and will return `nil` when the `[k]`
lookup returns `nil` _or_ when the key `k` is missing.

Then in `addPatchValueToDiff` we assert:
```go
contract.Assertf(v != nil || old != nil || oldInput != nil || newInput != nil)
```

but we don't have enough information to distinguish the "all values were
`nil`" case from the "one value was `nil` and the other was missing"
case. The assertion is only valid if all keys were present, hence we
panic when a key is removed when it previously had a value of `nil`.

The easiest fix is to relax the assertion (b886cfb8d) -- allow the
all-`nil` case since it can be valid. The downside is a poorer user
experience, because we don't have enough information to say whether the
update is an add or a delete; we could call it an update (odd) or a
no-op (lossy). The upside is it's a much smaller change with less risk.

Alternatively, we can have a better user experience by making
`addPatchValueToDiff` aware of whether keys were missing. This PR does
this by introducing a `missing` sentinel to distinguish from `nil`
values. This is riskier, but fortunately our test coverage is
comprehensive.

A failing unit test was added, along with another test to cover a small
bit of logic we weren't testing.

Fixes https://github.com/pulumi/pulumi-kubernetes/issues/1970.