Doc updates for overlay resources (#3064)
<!--Thanks for your contribution. See [CONTRIBUTING](CONTRIBUTING.md)
    for Pulumi's contribution guidelines.

    Help us merge your changes more quickly by adding more details such
    as labels, milestones, and reviewers.-->

### Proposed changes

<!--Give us a brief description of what you've done and what it solves.
-->
Some minor updates to the API docs for the overlay components:
- link to the newer version of the respective component, since the newer
versions have fixes for some of the long-standing issues with these
overlays.
- clarify the supported languages

I had considered adding a "known issues" section, e.g. to mention the
`dependsOn` limitation, but I felt it would be overkill, and that simply
linking to the newer version would suffice.

### Related issues (optional)

<!--Refer to related PRs or issues: #1234, or 'Fixes #1234' or 'Closes
#1234'.
Or link to full URLs to issues or pull requests in other GitHub
repositories. -->

Closes #1773
Closes #2181