Fix previews in SSA mode when there are replacements due to immutable fields (#3053)
### Proposed changes

This PR adds special handling for previews with SSA when we have a
replacement (delete before create) operation. There is no easy way to
tell the API server that the dry-run to be run is an object replacement.
As such, a workaround that we can do is to send a different object name
so the API server thinks we are creating a new object, so a already
exists type error doesn't occur.

Note, because of how SSA creates work currently with upserts, we'll
never face the `AlreadyExists` error, so we do not have to account for
that in this PR.

As an aside, `kubectl` also supports delete before create replacements
with the `replace` sub-command. A user would need to run `kubectl
replace --force -f .` to recreate any resources that update an immutable
field. `kubectl` does not support `-dry-run=server` with the `--force`
flag and the follow error is shown `--dry-run can not be used when
--force is set` since the API server does not support previewing this
type of operation.

### Related issues (optional)

Fixes: #2575