CustomResource should have plain apiVersion and kind properties (#3079)
<!--Thanks for your contribution. See [CONTRIBUTING](CONTRIBUTING.md)
    for Pulumi's contribution guidelines.

    Help us merge your changes more quickly by adding more details such
    as labels, milestones, and reviewers.-->

### Proposed changes

<!--Give us a brief description of what you've done and what it solves.
-->
This PR seeks to make the `apiVersion` and `kind` properties be
[plain](https://www.pulumi.com/docs/using-pulumi/pulumi-packages/schema/#type)
(prompt) values when configuring a `CustomResource`. A plain value is
required to eagerly compute the resource type token, e.g.
`kubernetes:stable.example.com/v1:CronTab` or
`kubernetes:stable.example.com/v1:CronTabPatch`.

Since the SDKs already accept an input, this PR strives to continue to
support it, rather than making a breaking change, if possible. That's
accomplished by synchronously waiting on the output data.

Three APIs are considered here:
1. creating or importing a CustomResource using resource args
2. getting an existing custom resource using the `get` helpers
3. patching a CustomResource using the "Patch" variant 

A summary of changes:
- Pulumi Schema - mark the `apiVersion` and `kind` properties to be
`plain: true`.
- Go SDK - Use `internals.UnsafeAwaitOutput` to synchronously resolve
the value.
- NodeJS SDK - Change to CustomResourceArgs to use plain property type
(BREAKING CHANGE)
- Java SDK - add some null checks

### Breaking Changes

#### NodeJS SDK

The args are retyped to `string` with expectation that most programs
will be unaffected.

Closes #1115 

### Related issues (optional)

<!--Refer to related PRs or issues: #1234, or 'Fixes #1234' or 'Closes
#1234'.
Or link to full URLs to issues or pull requests in other GitHub
repositories. -->