feat: enable converting CRD manifests to OpenAPI swagger
To be able to use the existing schemagen that the k8s provider uses, we'll need to convert any CRD manifest to their served OpenAPI swagger spec. We'll be relying on the existing stable schemagen logic of p-k and deprecating the schemagen logic in crd2pulumi.

chore: go mod tidy