Enhance detection logic for patch variants (#3102)
### Proposed Changes

This PR improves our Patch resource detection logic by removing the
reliance on a hardcoded list. Hardcoded lists are fragile because new
Patch resources introduced in future SDK versions, or those generated
dynamically by tools like `crd2pulumi`, would **not** be recognized as
Patch variants by our provider. This issue is already present with
`CustomResourcePatch` resources, as CustomResource kinds will never be
included in a hardcoded list.

The updated logic now compares the resource URN and the object's Kind.
Since the token within the URN of a patch variant always follows the
pattern `kind + "Patch"`, we can use this for comparison. Previously,
the logic simply checked if the URN ended in `Patch`, which was
inaccurate due to the possibility of a Custom Resource (CR) Kind ending
in `Patch` (e.g., "MyResourcePatch"). This would lead to the Pulumi
patch variant URN being `MyResourcePatchPatch`.

#### Testing

- Added integration tests for the YAML, Node.js, and Go languages since
CustomResources are overlays and not supported in YAML. Existing tests
did not ensure that deleted Patch resources only removed the fields
applied, and not the whole object. This behavior is now exercised as
well.
- Expanded existing unit tests.

Note: I did not expand the test coverage for Python, Java, and .NET at
this time, as porting the test programs for these languages requires
additional effort. We may revisit and add tests for these languages in
the future.

### Related Issues (Optional)

Fixes: #3023