fix: properly determine list objects in schemagen (#3205)
CRDs can have types that end in `List`. These are not actually List
resources, so we should not rely on name alone to determine this.
Instead, we should follow the rules set out in the [conventions
guide](https://github.com/kubernetes/community/blob/master/contributors/devel/sig-architecture/api-conventions.md#types-kinds)
to determine is a Kind is meant to be a List kind.