fix: ensure namespaces do not collide when generating canonicalGroup names
This commit ensures that we do not have package collisions when storing
canonical group names. Additional consideration is required to ensure
that there are no final schema changes so that we do not break v4 users.
It is not expected to have any package collisions in the p-k provider,
but is possible in the downstream crd2pulumi consure of our schemagen.
Here, users can supply any number of CRDs to typegen, so we need to be
specific when it comes to handling group names.