Fix await annotation precedence (#3344)
Currently when creating or updating, custom await annotations (including
`waitFor` and `skipAwait`) will be ignored for resources which have
built-in await logic.

This corrects the behavior by returning a boolean from
`metadata.ReadyCondition` indicating whether the user specified some
behavior via an annotation. If that's true then we don't use the
built-in logic.

We preserve the existing precedence when `PULUMI_K8S_AWAIT_ALL` is
`true`. That is, we will continue to use built-in await logic (if the
resource has it) by default, even when generic await is enabled.

During testing I also realized we have a bug where if a JSONPath is
specified pointing to a non-primitive type (like `.metadata`) it will
never resolve. I fixed this to match correctly.

Refs https://github.com/pulumi/pulumi-kubernetes/issues/3329
Fixes https://github.com/pulumi/pulumi-kubernetes/issues/3345