Normalize `x-kubernetes-*` fields for outputs (#3348)
### Proposed changes

This PR normalizes outputs so that any `x-kubernetes-*` fields are
stored in their respective underscored forms. When creating a CRD with
Pulumi, we are able to set the `x-kubernetes-*` fields correctly during
create/update (except in Java). However, if we attempt to read this
field as an output, it would fail since the stored state is not
normalized to the underscored forms. In other languages (Go/Python/TS),
this was fine - but in Java, this fails as the builder pattern we use
for codegen is unable to set the `x-kubernetes-*` fields since we only
have setters for the underscored versions.

This also fixes the issue where we cannot read the values of
`x_kubernetes_*` fields from outputs.

This PR adds additional unit tests, expands on the existing nodejs e2e
test to verify that we can now read `x_kubernetes_*` fields, and creates
a new Java e2e test to ensure we can successfully create CRDs that
contain `x_kubernetes_*` fields.

### Related issues (optional)

Closes: #3325