Break circular dependency between provider/tests (#3385)
The `./tests` modules currently depends on the `./provider` module for
some OpenAPI helpers, but `./provider` also depends on `./tests` for
Gomega matchers.

Since `./tests` isn't versioned like a proper module this puts us in a
situation where Renovate will continuously try to update it
https://github.com/pulumi/pulumi-kubernetes/pull/3384

We can teach Renovate to ignore this bump, or we can break the
dependency by moving some code around. I've opted for the latter, moving
`tests/gomega` to `provider/pkg/gomega`. (This makes it easier to
consume from other projects if we want to.)

Of course it's worth mentioning this is a self-inflicted problem due to
our module structure. Related
https://github.com/pulumi/ci-mgmt/pull/1238.