Support RFC9535 JSONPaths (#3658)
This is an alternative to #3634 which switches our JSONPath parser to the
excellent github.com/theory/jsonpath implementation of
[RFC9535](https://www.rfc-editor.org/rfc/rfc9535.html). This allows us to
support more complex conditions while still staying close to the existing
`kubectl` experience.

I prefer this approach because it's _possible_ `kubectl` will update their own
parser in the future to more closely match the spec. Inventing our own
syntax/conventions to describe boolean logic guarantees we will drift further
from `kubectl`, and it gives our users one more thing to learn. JSONPath is
hard enough to grok as it is :)

Note that value comparisons are still static, so something like
`{.spec.readyReplicas}={.spec.replicas}` is still not possible. If there's a
strong case for it we can re-consider adding it, but I'm skeptical for now.

Fixes https://github.com/pulumi/pulumi-kubernetes/issues/3521.