Fix a race condition during deletion (#3688)
If an informer is started after an object has been deleted we won't see
any events for it. This creates a race condition here:

https://github.com/pulumi/pulumi-kubernetes/blob/18a34579dec0083b7412f84e0f2dc2b0821756bd/provider/pkg/await/condition/deleted.go#L69-L80

This is a race because either (a) the goroutine observing events might
execute after `refreshClusterState` has run, or, more likely, (b) the
goroutine's informer can take a little while to get established due to
throttling/backoff/etc.

In both cases we have (1) a GET succeed with 200, (2) the object is
deleted on the cluster, (3) our informer finally starts watching for
events and never sees anything.

The repro in #3317 mentions it only happens on faster hardware, but I
suspect it's actually a side effect of that faster hardware running a
cluster under more load, making throttling more likely. That's my guess,
at least -- @aureq could weigh in.

To fix the issue I'm adding a new `DeletionSource` which resolves the
race by guaranteeing we always perform a quorum read after the informer
has synced. If we find the object has already been deleted we generate a
synthetic DELETED event.

I intentionally opted _not_ to touch the existing deleted condition's
logic because I want to make this fix as small as possible and I don't
want to risk inadvertently introducing other race conditions elsewhere.

Additionally I've only been able to occasionally reproduce the slowness
so this is missing a unit/regression test. I don't think it would be a
good use of time to try to create such a test. I've run the repro in
#3317 several hundred times with this fix and haven't observed the
slowness.

Fixes https://github.com/pulumi/pulumi-kubernetes/issues/3317