Preventing `$GOPATH/bin` to be created as file.
`pulumi-kubernetes` project Makefile targets depend on `$GOPATH` being
set and having the right directory structure. Since I was setting up
`mise`, I didn't have a global Golang installation present on my local
dev environment. Hence, `$GOPATH` and the directories that are usually
found on a Golang were not present.

When the following code was run as part of a Makefile target:

```sh
cp $(WORKING_DIR)/bin/${PROVIDER} ${GOPATH}/bin
```

It created a file called `bin` instead of a directory. This was causing
the integration tests to fail locally as it was unable to find the
required provider binaries.

This commit modifies the `cp` command to make explicit it expects bin to
be a directory.