Preventing `$GOPATH/bin` to be created as file. (#3719)
`pulumi-kubernetes` project Makefile targets depend on `$GOPATH` being
set and having the right directory structure. Since I was setting up
`mise`, I didn't have a global Golang installation present on my local
dev environment. Hence, `$GOPATH` and the directories that are usually
found on a Golang were not present.

When the following code was run as part of a Makefile target:

```sh
cp $(WORKING_DIR)/bin/${PROVIDER} ${GOPATH}/bin
```

It created a file called `bin` instead of a directory with the provider
binaries as expected. This was causing the integration tests to fail
locally as it was unable to find the required provider binaries.

This commit modifies the `cp` command to make explicit it expects bin to
be a directory.

<!--Thanks for your contribution. See [CONTRIBUTING](CONTRIBUTING.md)
    for Pulumi's contribution guidelines.

    Help us merge your changes more quickly by adding more details such
    as labels, milestones, and reviewers.-->

### Proposed changes

<!--Give us a brief description of what you've done and what it solves.
-->

### Related issues (optional)

<!--Refer to related PRs or issues: #1234, or 'Fixes #1234' or 'Closes
#1234'.
Or link to full URLs to issues or pull requests in other GitHub
repositories. -->