Update a redis image (#3917)
This is currently pointing at a v1 Docker manifest that's no longer
supported
on our GKS cluster:

> Failed to pull image "k8s.gcr.io/redis:e2e": Docker Image Format v1
> and Docker Image manifest version 2, schema 1 support has been
removed.
> Suggest the author of k8s.gcr.io/redis:e2e to upgrade the image to the
> OCI Format or Docker Image manifest v2, schema 2. More information at
> https://docs.docker.com/go/deprecated-image-specs/

This updates the image to one which is valid, and it fixes our failing
Dotnet test to no longer point at the contents of master.

Fixes https://github.com/pulumi/pulumi-kubernetes/issues/3904.